Working query param for amendments
