+++ app/templates/components/-main-map-layers.hbs
@@ -1,0 +1,21 @@
+{{!-- Bind layer groups' layers --}}
+{{#each visibleLayers key='id' as |layerObject|}}
+  {{mapbox-gl-on 'mousemove' layerObject.style.id (action 'handleLayerMouseMove') 
+    eventSource=map}}
+  {{mapbox-gl-on 'mouseleave' layerObject.style.id (action 'handleLayerMouseLeave') 
+    eventSource=map}}
+  {{mapbox-gl-on 'click' layerObject.style.id (action 'handleLayerMouseClick') 
+    eventSource=map}}
+  {{mapbox-gl-layer 
+      map=map 
+      layer=layerObject.style}}
+{{/each}}
+
+{{!-- Highlighted Layer Handling --}}
+{{#if highlightedFeature}}
+  {{mapbox-gl-layer 
+    map=map 
+    layer=highlightedFeatureLayer}}
+{{/if}}
+
+{{yield}}
